Fix a few bugs in the stack trie (#9)
* Fix a few bugs

* Fix the remaining bugs

* Move the stack code to its own file

* More PR grooming
